requirements.txt
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
torch==2.5.1+cu124
torchvision==0.20.1+cu124
